Organize code into folders. Update build config.
