+++ .gitignore
@@ -1,8 +1,8 @@
 .vscode
 %localappdata%
-ultralight*/
+/ultralight/
 System Volume Information
-src/*.obj
+src/**/*.obj
 demo/ultralight.log
 demo/default
 demo/inspector/
