+++ SConstruct
@@ -15,6 +15,8 @@ env = SConscript("godot-cpp/SConstruct")
 # tweak this if you want to use different folders, or more folders, to store your source code in.
 env.Append(CPPPATH=["src/"])
 sources = Glob("src/*.cpp")
+sources = sources + Glob("src/*/*.cpp")
+sources = sources + Glob("src/*/*/*.cpp")
 
 if env["platform"] == "macos":
     library = env.SharedLibrary(
