+++ src/app_example.hpp
@@ -1,7 +1,7 @@
 #ifndef APPEXAMPLE_H
 #define APPEXAMPLE_H
 
-#include "html_rect.hpp"
+#include "godot/html_rect/html_rect.hpp"
 
 namespace godot {
