+++ src/godot/inspector_rect/inspector_rect.cpp
