+++ src/godot/ultralight_manager/ultralight_manager.hpp
@@ -5,8 +5,8 @@
 #include <AppCore/Platform.h>
 #include <JavaScriptCore/JavaScript.h>
 
-#include "godot_file_system.hpp"
-#include "godot_clipboard.hpp"
+#include "ultralight/godot_file_system.hpp"
+#include "ultralight/godot_clipboard.hpp"
 
 using namespace ultralight;
