+++ src/godot/ultralight_singleton/ultralight_singleton.cpp
