+++ src/godot/view_rect/view_rect.cpp
@@ -1,6 +1,6 @@
 #include "view_rect.hpp"
 
-#include "ultralight_singleton.hpp"
+#include "godot/ultralight_singleton/ultralight_singleton.hpp"
 
 #include <godot_cpp/core/class_db.hpp>
 #include <godot_cpp/classes/rendering_server.hpp>
